Refactor get connection names (#168)
Refactor get connection names.

In our current implementation, we generate tools twice: once when
creating a Flow using `Flow.from_yaml`, and again when creating a
FlowExecutor and resolving the tool. To improve this, we have removed
the former step. However, this change breaks the logic of retrieving
connection names, as we still need to find the connection type input in
the tool. To address this issue, we have refactored the logic for
retrieving connection names that we can identify the connection type
input in the tool by simply loading the tool without complicated
resolving logics.

---------

Co-authored-by: Lina Tang <linatang@microsoft.com>