Handle JSONDecodeError exception when executor is resolving the inputs type convert (#544)
# Description

This is for bugfixing :

when user input List or Object, we use load user input as Json string
and parse it as json first. If the json string is not valid, then error
would raised. We used to raise InputTypeError, but this definitely not
illustrate the essence. So we try catch this specific exception for
special handling.

# All Promptflow Contribution checklist:
- [x] **The pull request does not introduce [breaking changes]**
- [x] **CHANGELOG is updated for new features, bug fixes or other
significant changes.**
- [x] **I have read the [contribution guidelines](../CONTRIBUTING.md).**

## General Guidelines and Best Practices
- [x] Title of the pull request is clear and informative.
- [ ] There are a small number of commits, each of which have an
informative message. This means that previously merged commits do not
appear in the history of the PR. For more information on cleaning up the
commits in your PR, [see this
page](https://github.com/Azure/azure-powershell/blob/master/documentation/development-docs/cleaning-up-commits.md).

### Testing Guidelines
- [ ] Pull request includes test coverage for the included changes.