[Executor] Add iterator output support (#576)
# Description

This pull request refactors the output check logic to use
`GeneratorProxy`. Previously, when the output was a `GeneratorType`, we
would wrap a proxy class that holds the generator as output. However,
this was an over condition, as we only need to wrap a proxy class on
iterator types as output. Therefore, we did the following changes:

- Replace `isinstance(output, GeneratorType)` with `isinstance(output,
Iterator)` on output wrapper logic.

This pull request fixes #561 

# All Promptflow Contribution checklist:
- [X] **The pull request does not introduce [breaking changes]**
- [X] **CHANGELOG is updated for new features, bug fixes or other
significant changes.**
- [X] **I have read the [contribution guidelines](../CONTRIBUTING.md).**

## General Guidelines and Best Practices
- [X] Title of the pull request is clear and informative.
- [X] There are a small number of commits, each of which have an
informative message. This means that previously merged commits do not
appear in the history of the PR. For more information on cleaning up the
commits in your PR, [see this
page](https://github.com/Azure/azure-powershell/blob/master/documentation/development-docs/cleaning-up-commits.md).

### Testing Guidelines
- [X] Pull request includes test coverage for the included changes.