[Feature] Improve error handling for loading tool error. (#556)
# Description

Currently, for loading tool error, the error message does not contain
the name of the node where the error occurred, which makes it difficult
for users to locate the problematic nod. So we need to improve the error
handling here.

Changes in this PR:

1. Define an error type: ToolLoadError, to encapsulate load errors
thrown from a third-party tool package. Its most important function is
to record the error target and module.
2. Define an error type: ResolveToolError, to append the name of the
failed node to the error message, which can help user locate the error
more quickly. ResolveToolError is a typeless error, so we need to rely
on inner_error to redefine its additional_info and error_codes.
3. To achieve the second point, we improved the handling of error_codes
in ExceptionPresenter.
4. Removes an unnecessary and ambiguous error type:
LoadToolError(ValidationException).
5. Supplement test cases.

# All Promptflow Contribution checklist:
- [ ] **The pull request does not introduce [breaking changes]**
- [x] **CHANGELOG is updated for new features, bug fixes or other
significant changes.**
- [ ] **I have read the [contribution guidelines](../CONTRIBUTING.md).**

## General Guidelines and Best Practices
- [x] Title of the pull request is clear and informative.
- [x] There are a small number of commits, each of which have an
informative message. This means that previously merged commits do not
appear in the history of the PR. For more information on cleaning up the
commits in your PR, [see this
page](https://github.com/Azure/azure-powershell/blob/master/documentation/development-docs/cleaning-up-commits.md).

### Testing Guidelines
- [x] Pull request includes test coverage for the included changes.