Enable node level tracing of generator output (#702)
# Description

This PR enables tracing of generator output of nodes in promptflow.
Previously, the `Trace` object was serialized as a JSON string in
`RunInfo` right before node execution ends. However, this did not
capture the generator output as it was not consumed yet. To fix this
issue, this PR makes the following changes:

- Instead of generating a JSON string in `RunInfo`, we track a list of
generator output in `RunInfo`, which can be updated when the generator
is consumed.
- We update the node `RunInfo` before the flow ends, at that time, all
generators are expected to be consumed.

This change will enable generator tracing in node level, which is
currently not possible in the Promptflow implementation.

# All Promptflow Contribution checklist:
- [X] **The pull request does not introduce [breaking changes].**
- [X] **CHANGELOG is updated for new features, bug fixes or other
significant changes.**
- [X] **I have read the [contribution guidelines](../CONTRIBUTING.md).**
- [X] **Create an issue and link to the pull request to get dedicated
review from promptflow team. Learn more: [suggested
workflow](../CONTRIBUTING.md#suggested-workflow).**

## General Guidelines and Best Practices
- [X] Title of the pull request is clear and informative.
- [X] There are a small number of commits, each of which have an
informative message. This means that previously merged commits do not
appear in the history of the PR. For more information on cleaning up the
commits in your PR, [see this
page](https://github.com/Azure/azure-powershell/blob/master/documentation/development-docs/cleaning-up-commits.md).

### Testing Guidelines
- [X] Pull request includes test coverage for the included changes.