[Open_Source_LLM] Update Tool for Ignite - LLaMa2 Improvements (#751)
# Description

This Pull Request improves on the success of the Open_Source_LLM with
quality of life improvements and a simplified interface. More details:

1. Support providing an endpoint_name in addition to the current
offering of a selecting a connection from dropdown. This text box is a
temporary option, until the dynamic list functionality is available.
2. Move the priority LLaMa2 model parameters to the 1st class citizens
on the Tool UX
3. Move key advanced LLaMa2 model parameters to on the Tool UX, under
the Advanced Parameters section.
4. Update tests to match this experience.
5. Clean up previous code.

Here is the current experience in the VsCode experience:

![image](https://github.com/microsoft/promptflow/assets/99283778/c56cd2b3-7c30-4119-981d-78da7d4a258a)

# All Promptflow Contribution checklist:
- [x] **The pull request does not introduce [breaking changes].**
- [x] **CHANGELOG is updated for new features, bug fixes or other
significant changes.**
- [x] **I have read the [contribution guidelines](../CONTRIBUTING.md).**
- [x] **Create an issue and link to the pull request to get dedicated
review from promptflow team. Learn more: [suggested
workflow](../CONTRIBUTING.md#suggested-workflow).**

## General Guidelines and Best Practices
- [x] Title of the pull request is clear and informative.
- [x] There are a small number of commits, each of which have an
informative message. This means that previously merged commits do not
appear in the history of the PR. For more information on cleaning up the
commits in your PR, [see this
page](https://github.com/Azure/azure-powershell/blob/master/documentation/development-docs/cleaning-up-commits.md).

### Testing Guidelines
- [x] Pull request includes test coverage for the included changes.

---------

Co-authored-by: Gerard <GerardJWoods@live.com>
Co-authored-by: chjinche <49483542+chjinche@users.noreply.github.com>