Change function call type to object (#815)
# Description

**Issue:**
The function_call can be of type 'str' or 'dict'. However, when
function_call is set to 'str' type in the char API, it becomes unsafe to
pass the parameter. Exceptions could occur when trying to convert a
'dict' input to a 'str' input in the process of passing this parameter.

**Solution:**
The type of function_call was changed from 'str' to 'object' to enhance
safety. The 'object' type is more versatile and can handle a wider range
of data types, thus reducing the risk of exceptions.

**Tests:** 
Additional unit and end-to-end tests were added to ensure the
functionality and safety of the changes. All tests were successfully
passed, indicating that the solution is effective and doesn't introduce
new issues.

# All Promptflow Contribution checklist:
- [X] **The pull request does not introduce [breaking changes].**
- [ ] **CHANGELOG is updated for new features, bug fixes or other
significant changes.**
- [X] **I have read the [contribution guidelines](../CONTRIBUTING.md).**
- [ ] **Create an issue and link to the pull request to get dedicated
review from promptflow team. Learn more: [suggested
workflow](../CONTRIBUTING.md#suggested-workflow).**

## General Guidelines and Best Practices
- [X] Title of the pull request is clear and informative.
- [X] There are a small number of commits, each of which have an
informative message. This means that previously merged commits do not
appear in the history of the PR. For more information on cleaning up the
commits in your PR, [see this
page](https://github.com/Azure/azure-powershell/blob/master/documentation/development-docs/cleaning-up-commits.md).

### Testing Guidelines
- [x] Pull request includes test coverage for the included changes.

---------

Co-authored-by: cs_lucky <si.chen@microsoft.com>