[Executor] Support ToolProvider for script tool (#560)
# Description

In this pull request, we have added support for `ToolProvider` to the
script tool. This allows users to initialize their script tool before
running it. With this update, users can define a script tool using a
class method and initialize the tool in the *\_\_init\_\_()* method of
the class.

This is a sample script tool:

```python
from promptflow import ToolProvider, tool


class ScriptToolWithInit(ToolProvider):
    def __init__(self, init_input: str):
        # This will only be called once when init the flow
        super().__init__()
        self.init_input = init_input  

    @tool
    def call(self, input: str):
        # This will be called multiple times
        return str.join(" ", [self.init_input, input])
```

# All Promptflow Contribution checklist:
- [x] **The pull request does not introduce [breaking changes]**
- [x] **CHANGELOG is updated for new features, bug fixes or other
significant changes.**
- [x] **I have read the [contribution guidelines](../CONTRIBUTING.md).**

## General Guidelines and Best Practices
- [x] Title of the pull request is clear and informative.
- [x] There are a small number of commits, each of which have an
informative message. This means that previously merged commits do not
appear in the history of the PR. For more information on cleaning up the
commits in your PR, [see this
page](https://github.com/Azure/azure-powershell/blob/master/documentation/development-docs/cleaning-up-commits.md).

### Testing Guidelines
- [x] Pull request includes test coverage for the included changes.

---------

Co-authored-by: Lina Tang <linatang@microsoft.com>