Assign promptflow-execution as Code Owner for Specific Directories (#1350)
# Description

This pull request introduces changes to the CODEOWNERS file to designate
promptflow-execution as the required code reviewer for specific folders
within the repository. This change aims to streamline the review process
by automatically involving the promptflow-execution team when changes
are made to the areas they own. By enforcing this, we ensure that any
modifications to these parts of the codebase are thoroughly reviewed by
the appropriate experts.

The folders for which promptflow-execution has been added as code owner
are:
- src/promptflow/promptflow/_core
- src/promptflow/promptflow/_internal
- src/promptflow/promptflow/batch
- src/promptflow/promptflow/executor
- src/promptflow/promptflow/integrations
- src/promptflow/promptflow/storage
- src/promptflow/tests/executor

# All Promptflow Contribution checklist:
- [X] **The pull request does not introduce [breaking changes].**
- [X] **CHANGELOG is updated for new features, bug fixes or other
significant changes.**
- [X] **I have read the [contribution guidelines](../CONTRIBUTING.md).**
- [X] **Create an issue and link to the pull request to get dedicated
review from promptflow team. Learn more: [suggested
workflow](../CONTRIBUTING.md#suggested-workflow).**

## General Guidelines and Best Practices
- [X] Title of the pull request is clear and informative.
- [X] There are a small number of commits, each of which have an
informative message. This means that previously merged commits do not
appear in the history of the PR. For more information on cleaning up the
commits in your PR, [see this
page](https://github.com/Azure/azure-powershell/blob/master/documentation/development-docs/cleaning-up-commits.md).

### Testing Guidelines
- [X] Pull request includes test coverage for the included changes.