add tool owner (#1530)
# Description

This pull request introduces changes to the CODEOWNERS file to designate
[promptflow-tool](https://github.com/orgs/microsoft/teams/promptflow-tools)
as the required code reviewer for specific folders within the
repository. This change aims to streamline the review process by
automatically involving the
[promptflow-tool](https://github.com/orgs/microsoft/teams/promptflow-tools)
team when changes are made to the areas they own. By enforcing this, we
ensure that any modifications to these parts of the codebase are
thoroughly reviewed by the appropriate experts.

The folders for which promptflow-tool has been added as code owner are:

/docs/how-to-guides/develop-a-tool
/src/promptflow-tools
/examples/tools
/scripts/tool

# All Promptflow Contribution checklist:
- [x] **The pull request does not introduce [breaking changes].**
- [x] **CHANGELOG is updated for new features, bug fixes or other
significant changes.**
- [x] **I have read the [contribution guidelines](../CONTRIBUTING.md).**
- [x] **Create an issue and link to the pull request to get dedicated
review from promptflow team. Learn more: [suggested
workflow](../CONTRIBUTING.md#suggested-workflow).**

## General Guidelines and Best Practices
- [ ] Title of the pull request is clear and informative.
- [ ] There are a small number of commits, each of which have an
informative message. This means that previously merged commits do not
appear in the history of the PR. For more information on cleaning up the
commits in your PR, [see this
page](https://github.com/Azure/azure-powershell/blob/master/documentation/development-docs/cleaning-up-commits.md).

### Testing Guidelines
- [ ] Pull request includes test coverage for the included changes.

---------

Co-authored-by: Yuhang Liu <yuhaliu@microsoft.com>