improve error category (#1753)
# Description
There are many error categories that are UKNOW, because the throw errors
are not of the `UserErrorException `or `SystemErrorException `typies, so
need to optimized.

![image](https://github.com/microsoft/promptflow/assets/23182548/93614bd3-320d-44fb-86bf-486159a9cd75)

Based on the data, I optimized the data with more anomalies:

![image](https://github.com/microsoft/promptflow/assets/23182548/afddc5d6-148f-4003-8ef3-8b0196794173)


Please add an informative description that covers that changes made by
the pull request and link all relevant issues.

# All Promptflow Contribution checklist:
- [x] **The pull request does not introduce [breaking changes].**
- [ ] **CHANGELOG is updated for new features, bug fixes or other
significant changes.**
- [ ] **I have read the [contribution guidelines](../CONTRIBUTING.md).**
- [ ] **Create an issue and link to the pull request to get dedicated
review from promptflow team. Learn more: [suggested
workflow](../CONTRIBUTING.md#suggested-workflow).**

## General Guidelines and Best Practices
- [x] Title of the pull request is clear and informative.
- [ ] There are a small number of commits, each of which have an
informative message. This means that previously merged commits do not
appear in the history of the PR. For more information on cleaning up the
commits in your PR, [see this
page](https://github.com/Azure/azure-powershell/blob/master/documentation/development-docs/cleaning-up-commits.md).

### Testing Guidelines
- [ ] Pull request includes test coverage for the included changes.