Remove unused __wrapped_function (#1803)
# Description

Currently, when a function is wrapped with `trace()`, the function will
set an attribute `__wrapped_function` to itself to note which function
has wrapped it.
However, this operation would fail when the function is a built-in
function. i.e.

```python
from promptflow import trace
traced_sum = trace(sum)
```

will get this error:
```
AttributeError: 'builtin_function_or_method' object has no attribute '__wrapped_function'
```

This error is because built-in functions in Python are implemented in C
and do not have a `__dict__` attribute where we can set arbitrary
attributes.

This attribute is not used in our code, and this PR is to remove it.

# All Promptflow Contribution checklist:
- [x] **The pull request does not introduce [breaking changes].**
- [ ] **CHANGELOG is updated for new features, bug fixes or other
significant changes.**
- [x] **I have read the [contribution guidelines](../CONTRIBUTING.md).**
- [ ] **Create an issue and link to the pull request to get dedicated
review from promptflow team. Learn more: [suggested
workflow](../CONTRIBUTING.md#suggested-workflow).**

## General Guidelines and Best Practices
- [x] Title of the pull request is clear and informative.
- [x] There are a small number of commits, each of which have an
informative message. This means that previously merged commits do not
appear in the history of the PR. For more information on cleaning up the
commits in your PR, [see this
page](https://github.com/Azure/azure-powershell/blob/master/documentation/development-docs/cleaning-up-commits.md).

### Testing Guidelines
- [x] Pull request includes test coverage for the included changes.