[Executor] Support new entry format for eager flow (#1870)
# Description
This pull request includes changes to the `promptflow` package,
primarily focusing on simplifying the handling of eager flow files. The
most significant changes include the removal of the `entry` parameter
from several methods and classes, and the modification of the flow file
loading process.

Changes to the flow file handling:

*
[`src/promptflow/promptflow/executor/_script_executor.py`](diffhunk://#diff-3294a0ccdf18c47385db91932bd860940cc3df5df10d206138fbcd48075530d5L28-L55):
The `entry` parameter was removed from the `ScriptExecutor` class and
its `__init__` method. The method `_load_func` was added to load the
function from the flow file, and the method `_parse_flow_file` was added
to parse the flow file and return the module name and function name.
[[1]](diffhunk://#diff-3294a0ccdf18c47385db91932bd860940cc3df5df10d206138fbcd48075530d5L28-L55)
[[2]](diffhunk://#diff-3294a0ccdf18c47385db91932bd860940cc3df5df10d206138fbcd48075530d5R104-R136)

*
[`src/promptflow/promptflow/executor/flow_executor.py`](diffhunk://#diff-faa6c81d614b7e41b18a42a93139d961d92afa9aa9dd0b72cb6b7176d7541e69L97):
The `entry` parameter was removed from the `FlowExecutor` class and its
`__init__` method. The methods `_is_eager_flow_yaml` and
`_create_from_flow` were modified to handle eager flow files without the
`entry` parameter, and the method `_parse_eager_flow_yaml` was removed.
[[1]](diffhunk://#diff-faa6c81d614b7e41b18a42a93139d961d92afa9aa9dd0b72cb6b7176d7541e69L97)
[[2]](diffhunk://#diff-faa6c81d614b7e41b18a42a93139d961d92afa9aa9dd0b72cb6b7176d7541e69L235)
[[3]](diffhunk://#diff-faa6c81d614b7e41b18a42a93139d961d92afa9aa9dd0b72cb6b7176d7541e69L271)
[[4]](diffhunk://#diff-faa6c81d614b7e41b18a42a93139d961d92afa9aa9dd0b72cb6b7176d7541e69L282-L298)

*
[`src/promptflow/promptflow/batch/_batch_engine.py`](diffhunk://#diff-ecf0905f2116abe08b5cf2931b856bf39aa8b38a30fadd9042538d076dbfde80L81):
The `entry` parameter was removed from the `BatchEngine` class and its
`__init__` method. The method `_is_eager_flow_yaml` was modified to
handle eager flow files without the `entry` parameter, and the method
`_parse_eager_flow_yaml` was removed.
[[1]](diffhunk://#diff-ecf0905f2116abe08b5cf2931b856bf39aa8b38a30fadd9042538d076dbfde80L81)
[[2]](diffhunk://#diff-ecf0905f2116abe08b5cf2931b856bf39aa8b38a30fadd9042538d076dbfde80L104-L118)
[[3]](diffhunk://#diff-ecf0905f2116abe08b5cf2931b856bf39aa8b38a30fadd9042538d076dbfde80L406-L420)

Removal of the `InvalidFlowFileError` class:

*
[`src/promptflow/promptflow/executor/_errors.py`](diffhunk://#diff-1f02407850413baccbb33e0078d9772e15ae1c46c020dc7457529c6f7a054315L279-L282):
The `InvalidFlowFileError` class was removed.

*
[`src/promptflow/promptflow/executor/flow_executor.py`](diffhunk://#diff-faa6c81d614b7e41b18a42a93139d961d92afa9aa9dd0b72cb6b7176d7541e69L47):
The import of the `InvalidFlowFileError` class was removed.

*
[`src/promptflow/promptflow/batch/_batch_engine.py`](diffhunk://#diff-ecf0905f2116abe08b5cf2931b856bf39aa8b38a30fadd9042538d076dbfde80L41):
The import of the `InvalidFlowFileError` class was removed.

Other changes:

*
[`src/promptflow/promptflow/executor/_line_execution_process_pool.py`](diffhunk://#diff-c14422c33faa9e44145738152120929455da4e4f05d4fc7b59158423fa609bc1L130):
The `entry` parameter was removed from the `__init__` method and the
`create_executor_fork` method.
[[1]](diffhunk://#diff-c14422c33faa9e44145738152120929455da4e4f05d4fc7b59158423fa609bc1L130)
[[2]](diffhunk://#diff-c14422c33faa9e44145738152120929455da4e4f05d4fc7b59158423fa609bc1L636)

*
[`src/promptflow/promptflow/batch/_python_executor_proxy.py`](diffhunk://#diff-8c42522c7186705de9a6651343af6006be02b7342f72467c2785f23f93ea669eL32-R36):
The `entry` parameter was removed from the `create` method.

*
[`src/promptflow/tests/executor/e2etests/test_eager_flow.py`](diffhunk://#diff-0f88413b3a6799b404c8406e28d7e95a18a99771f78a65cabaaf2ac5e0a6ec8bR49):
Several tests were skipped due to the changes in the handling of eager
flow files.
[[1]](diffhunk://#diff-0f88413b3a6799b404c8406e28d7e95a18a99771f78a65cabaaf2ac5e0a6ec8bR49)
[[2]](diffhunk://#diff-0f88413b3a6799b404c8406e28d7e95a18a99771f78a65cabaaf2ac5e0a6ec8bL180-L182)

*
[`src/promptflow/tests/sdk_cli_test/e2etests/test_flow_test.py`](diffhunk://#diff-acf3c2285ac2a7051d59e3a69abe02224e2d6d5d69e3a0aabd2afd9311f87657R245-R257):
Several tests were skipped due to the changes in the handling of eager
flow files.

*
[`src/promptflow/tests/sdk_cli_test/e2etests/test_flow_run.py`](diffhunk://#diff-94a59a05643476869fa3c6bc45466f1582944a935488075e2e63b6a6a196958fR1241):
One test was skipped due to the changes in the handling of eager flow
files.

*
[`src/promptflow/tests/test_configs/eager_flows/dummy_flow_with_exception/flow.dag.yaml`](diffhunk://#diff-df24e058d75dc22a54e4c3ab6f05323adac4896d40e3967bedb332aaf45cdbceL1-R1):
The `path` and `entry` keys were replaced with a single `entry` key
containing the module name and function name.

# All Promptflow Contribution checklist:
- [x] **The pull request does not introduce [breaking changes].**
- [ ] **CHANGELOG is updated for new features, bug fixes or other
significant changes.**
- [x] **I have read the [contribution guidelines](../CONTRIBUTING.md).**
- [ ] **Create an issue and link to the pull request to get dedicated
review from promptflow team. Learn more: [suggested
workflow](../CONTRIBUTING.md#suggested-workflow).**

## General Guidelines and Best Practices
- [x] Title of the pull request is clear and informative.
- [x] There are a small number of commits, each of which have an
informative message. This means that previously merged commits do not
appear in the history of the PR. For more information on cleaning up the
commits in your PR, [see this
page](https://github.com/Azure/azure-powershell/blob/master/documentation/development-docs/cleaning-up-commits.md).

### Testing Guidelines
- [x] Pull request includes test coverage for the included changes.

---------

Co-authored-by: Lina Tang <linatang@microsoft.com>