[doc] Add doc for automatic runtime (#1952)
# Description

This pull request primarily focuses on updating the documentation and
making a minor change to the test suite. The most significant changes
include the addition of a new guide on creating runs with automatic
runtime for Azure AI, an update to the column mapping guide, and a
change in the import statement in the test suite.

Documentation updates:

*
[`docs/cloud/azureai/create-run-with-automatic-runtime.md`](diffhunk://#diff-6e1bdeb922a4368671b33de9706d1e8e25714cf26402c37ae172f91c287e6da4R1-R108):
A new guide was added that provides detailed instructions on how to
create a run with automatic runtime in Azure AI. This includes how to
specify pip requirements, customize the automatic runtime, and execute
the run using both the CLI and SDK.
*
[`docs/cloud/index.md`](diffhunk://#diff-06412129710f6abd0bb3b3efb0befb0225bad060c79551996d54192fb0573037R36):
The new guide on creating runs with automatic runtime was added to the
list of Azure AI resources.
*
[`docs/how-to-guides/run-and-evaluate-a-flow/use-column-mapping.md`](diffhunk://#diff-0c6bd53486b1a7c57438fb7e4498e4568154b9b99d778cd3d72b8ae71850e07cR12):
The guide was updated to include `${run.inputs.<input_name>}` as a
supported value in column mapping when `--run` is provided for `pf run`.

Test suite change:

*
[`src/promptflow/tests/sdk_cli_azure_test/unittests/test_flow_entity.py`](diffhunk://#diff-44e987aed05f781a28e90e24fda551a5391751a1d4f59af03a8c1c05cb058720L13-R13):
The `load_run` function was imported directly from `promptflow` instead
of `promptflow._sdk._load_functions`, simplifying the import statement.
# All Promptflow Contribution checklist:
- [ ] **The pull request does not introduce [breaking changes].**
- [ ] **CHANGELOG is updated for new features, bug fixes or other
significant changes.**
- [ ] **I have read the [contribution guidelines](../CONTRIBUTING.md).**
- [ ] **Create an issue and link to the pull request to get dedicated
review from promptflow team. Learn more: [suggested
workflow](../CONTRIBUTING.md#suggested-workflow).**

## General Guidelines and Best Practices
- [ ] Title of the pull request is clear and informative.
- [ ] There are a small number of commits, each of which have an
informative message. This means that previously merged commits do not
appear in the history of the PR. For more information on cleaning up the
commits in your PR, [see this
page](https://github.com/Azure/azure-powershell/blob/master/documentation/development-docs/cleaning-up-commits.md).

### Testing Guidelines
- [ ] Pull request includes test coverage for the included changes.