[promptflow][internal] Fix `cumulative_token_count` schema error (#2049)
# Description

There is a schema bug for `cumulative_token_count`, it should be an
optional dictionary, instead of an optional string.

This PR targets to fix this, so that we can have correct response, and
it will be correctly rendered in trace UI; also try to add an end-to-end
test to guard this behavior.

This PR also fixes missing update for MANIFEST in #2035 , which leads to
missing files when build.

# All Promptflow Contribution checklist:
- [x] **The pull request does not introduce [breaking changes].**
- [ ] **CHANGELOG is updated for new features, bug fixes or other
significant changes.**
- [x] **I have read the [contribution guidelines](../CONTRIBUTING.md).**
- [ ] **Create an issue and link to the pull request to get dedicated
review from promptflow team. Learn more: [suggested
workflow](../CONTRIBUTING.md#suggested-workflow).**

## General Guidelines and Best Practices
- [x] Title of the pull request is clear and informative.
- [x] There are a small number of commits, each of which have an
informative message. This means that previously merged commits do not
appear in the history of the PR. For more information on cleaning up the
commits in your PR, [see this
page](https://github.com/Azure/azure-powershell/blob/master/documentation/development-docs/cleaning-up-commits.md).

### Testing Guidelines
- [x] Pull request includes test coverage for the included changes.